requirements.txt
Django==1.6.1
South==0.8.4
Werkzeug==0.9.4
dj-database-url==0.2.2
django-cors-headers==0.12
django-rq==0.6.1
django-uuidfield==0.4.0
django-werkzeug-debugger-runserver==0.2
djangorestframework==2.3.10
envoy==0.0.2
gunicorn==18.0
jsonfield==0.9.20
oauthlib==0.6.0
psycopg2==2.5.2
python-dateutil==2.2
pytz==2013.9
rauth==0.6.2
redis==2.9.1
requests==2.1.0
requests-oauthlib==0.4.0
rq==0.3.13
shapeways==0.1.6
six==1.4.1
times==0.6.2
wsgiref==0.1.2
